Finalise mode switching
Added mode switching to Master and Interceptor, fixed the obvious bugs in the App resulting from this.

Still needs tidying up and testing in the car but this looks very promising at the moment.

+++ App/OpenHaldex/app/src/main/java/com/kong/openhaldex/LockPoint.java
@@ -3,13 +3,13 @@ package com.kong.openhaldex;
 import java.io.Serializable;
 
 public class LockPoint implements Serializable {
-    float speed;
-    float lock;
-    float intensity;
+    int speed;
+    int lock;
+    int intensity;
 
     public LockPoint() {}
 
-    public LockPoint(float _speed, float _lock, float _intensity){
+    public LockPoint(int _speed, int _lock, int _intensity){
         speed = _speed;
         lock = _lock;
         intensity = _intensity;
